Implements the mouse commands for webdriver.
Move the mouse over an element.
/session/:sessionId/element/:id/hover

Drag and drop an element. The distance to drag an element should be
specified relative to the upper-left corner of the page.
/session/:sessionId/element/:id/drag

Click on an element.
/session/:sessionId/element/:id/click

The events in this CL use Webkit events and do not use OS level calls.

BUG=none
TEST=webdriver_remote_tests.py

Review URL: http://codereview.chromium.org/4681007

git-svn-id: http://src.chromium.org/svn/trunk/src@75618 4ff67af0-8c30-449e-8e8b-ad334ec8d88c