Recreate both GpuProcessHost and GpuProcessHostUIShim when a new GPU process is launched.
This is so that a renderer that was previously sending messages to the browser regarding a particular GPU process, it can be informed that a new one has been launched and that it has to start over. This part is not done yet.

Both GpuProcessHost and GpuProcessHostUIShim are now identified by a common host ID. Neither are singletons.

This is also a step towards allowing multiple GPU processes.

TEST=trybots, check WebGL and 3D CSS work, check terminating Chrome works while running WebGL, check --single-process works, checking terminated GPU process is relaunched and that 3D CSS restarts automatically
BUG=none

Review URL: http://codereview.chromium.org/6473003

git-svn-id: http://src.chromium.org/svn/trunk/src@75645 4ff67af0-8c30-449e-8e8b-ad334ec8d88c