Disable the resize IPC for accelerated compositing on TOUCH_UI.
The rendering path for TOUCH_UI is a work in progress, but it will roughly resemble the MAC OS X path. In particular, we will not need the browser to synchronously resize the window that the GPU process is drawing into. This patch makes accelerated rendering closer to usable on TOUCH_UI (the chrome is distorted).

BUG=none (accelerated compositing for TOUCH_UI is still NOT_IMPLEMENTED)
TEST=by hand on Linux (w/ and w/o TOOLKIT_VIEWS and TOUCH_UI), Windows, and Mac

Review URL: http://codereview.chromium.org/6567005

git-svn-id: http://src.chromium.org/svn/trunk/src@75884 4ff67af0-8c30-449e-8e8b-ad334ec8d88c