Fix up mac test client and simple host for remoting.
Remoting's test client and simple host weren't initializing the NSApplication correctly which is required for MessageLoops to work on the Mac. These are just test apps that we use, so I felt it was ok to bring in code from base/test.

Also added a check to message_pump_mac.mm to verify that NSApp is initialized before we attempt to use it.

BUG=NONE
TEST=BUILD

Review URL: http://codereview.chromium.org/6461023

git-svn-id: http://src.chromium.org/svn/trunk/src@76624 4ff67af0-8c30-449e-8e8b-ad334ec8d88c