Avoid holding a lock during destructor call when not needed
I noticed this while reviewing related code.  I don't
know of an actual bug, but it is bad form to hold a
lock when calling a function (destructor) unless
necessary.  For example, if the destructor posted
a message, it might cause a hang on linux as the
posting could re-enter this helper.  It is also
possible that the task deletion will take an
extended period of time, and that can block
other concurrent posts from taking place.

r=darin

Review URL: http://codereview.chromium.org/6602024

git-svn-id: http://src.chromium.org/svn/trunk/src@76774 4ff67af0-8c30-449e-8e8b-ad334ec8d88c