Fix URLFetcher::CancelAll().
URLFetcher::CancelAll() has re-entrant behavior. We swap all the fetchers out of the global registry and then try to cancel them. This causes a problem because cancellation will try to update the global registry. This fixes the problem by not swapping the fetchers. Instead, we iterate until the registry is empty. This also has the property of working when killing the URLRequestJobs may recursively invoke behavior that creates/removes URLFetchers.

Adds a new test for this.

BUG=75212
TEST=URLFetcherTest.CancelAll

Review URL: http://codereview.chromium.org/6627072

git-svn-id: http://src.chromium.org/svn/trunk/src@77346 4ff67af0-8c30-449e-8e8b-ad334ec8d88c