Add ability to find out whether an extension pref value is coming from incognito preferences.
If the incognito flag in the request details is set, we now return an incognito parameter from experimental.preference.get that specifies whether the value is coming from the incognito preferences or the regular ones.

Also, return an error if an extension that isn't enabled in incognito mode is trying to access incognito preferences.

BUG=73994
TEST=ExtensionApiTest.*

Review URL: http://codereview.chromium.org/6628081

git-svn-id: http://src.chromium.org/svn/trunk/src@77465 4ff67af0-8c30-449e-8e8b-ad334ec8d88c