Adding an image to avoid future try server failure.
This is for the upcoming redesign changelist for the
redesign of the Manage Extensions page (as a Settings
page).

Try servers tend to fail (at least have in the past)
if a binary file was included in the changelist
being 'tried'.

This should avoid that problem.

BUG=52447
TEST=None
TBR=mwichary

Review URL: http://codereview.chromium.org/6652021

git-svn-id: http://src.chromium.org/svn/trunk/src@77471 4ff67af0-8c30-449e-8e8b-ad334ec8d88c