Ensure that PP_Instance values are unique within a plugin process in addition to within the renderer.
This works by having the renderer check with the plugin that a PP_Instance is
available before using it. If it's already seen, the renderer will generate a
new PP_Instance and retry.

For performance, this message is handled on the I/O thread of the plugin so it
will not be blocked by the actual plugin code.

This requires an unfortunate amount of plumbing. Since the renderer can't
depend directly on the proxy, we have a new PPB_Proxy function to set the
verification function used to perform this check.

There is also a new plugin dispatcher delegate where I moved some of the global
state to that used to go into the init function. Adding yet another parameter
there seemed unfortunate.

TEST=manual
BUG=74961
Review URL: http://codereview.chromium.org/6628019

git-svn-id: http://src.chromium.org/svn/trunk/src@77693 4ff67af0-8c30-449e-8e8b-ad334ec8d88c