Show details about allowed and blocked cookies in a tabbed pane.  Also reworks the UI to show allowed and blocked cookies in differnt tabs.  This implement only addresses the windows build.
BUG=63650
TEST=Block all cookies, and visit a site that uses cookies.  Click on the broken
cookie icon that appears in the URL bar, and then click to see the cookies
for this page.  The UI now shows the allowed and blocked cookies in separate
tabs, with an area below showing information about specific cookies selected
in the tree control.

Review URL: http://codereview.chromium.org/6639013

git-svn-id: http://src.chromium.org/svn/trunk/src@77851 4ff67af0-8c30-449e-8e8b-ad334ec8d88c