Add redundancy to detect double frees in TCMalloc
Added a single byte or word (tranparently) to all
allocations, and wrote a unique value into that
location after each allocation. When free() is
called, we validate the flag, and then mark
the block as not being allocated.

Any time a block fails to validate, we crash,
as this means the object either overran its
allocated region (or suffered memory corruption),
or else (more likely) a double free took place.
We have two distinct crash stacks for the
two distinct validation problems (corrupt vs
double free).

I did a first landing of this patch to test perf impact
and see what corruption it surfaced in:
http://src.chromium.org/viewvc/chrome?view=rev&revision=77941
It appears that Windows is not using TCMalloc on our bots.



TBR=mbelshe

Review URL: http://codereview.chromium.org/6683027

git-svn-id: http://src.chromium.org/svn/trunk/src@77948 4ff67af0-8c30-449e-8e8b-ad334ec8d88c