Clean up temporary files and folders left over from previous installations.
Installation failures due to out-of-disk-space seem to be getting more
and more frequent and upon inspection, I found that we can in some cases
do a terrible job and leave cruft behind.  Worst of all we can do this if
an installation fails due to low free disk space, thus only adding to the
problem.  I suspect there maybe users stuck on a specific revision of
Chrome because of this.

I'm also implementing Robert's idea of supporting for a --cleanup command
line arg that can be passed to the mini installer to only do the cleanup
and not the installation.

BUG=74777
TEST=This should reduce the number of crashes we're getting due to out of disk space.

Review URL: http://codereview.chromium.org/6670023

git-svn-id: http://src.chromium.org/svn/trunk/src@78424 4ff67af0-8c30-449e-8e8b-ad334ec8d88c