Videos on cnn.com would not paint at times especially if a number of them were opened simultaneously.
This CL fixes problem which I was seeing on my setup which makes it a touch better. 

However the videos still fail to load at times. The interesting aspect is when the video fails
to load the other widgets on the page also don't paint.

These videos are all implemented as flash windowless plugins. When a flash plugin is instantiated
it invokes NPN sync calls like NPN_Invoke etc in the context of NPP_New. The renderer in the
meantime attempts to retrieve cookies from the browser via sync IPC messages. we set the
pump messages flag to true in this context. As a result we end up in a scenario where the
plugin in the context of the NPN_Invoke call mentioned above starts pumping messages resulting
in other plugin instances getting created in this context resulting in flash getting confused.

Fix is to only pump the cookie messages for chrome frame processes where we are at the mercy
of the cookie policy in the host.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=71591

BUG=71591
TEST=manually as described in the bug at this point.

Review URL: http://codereview.chromium.org/6686020

git-svn-id: http://src.chromium.org/svn/trunk/src@78463 4ff67af0-8c30-449e-8e8b-ad334ec8d88c