Provisional fix for WebFrame API change upstream (webkit bug 56139).
This is somewhat hackish.. the new version of the function upstream has a default initializer. I think it's better though to add it this way to remove ambiguity - for the overloading the default initializer should not matter.

Either way, I expect to remove the obsolete definition within a day or so.

Review URL: http://codereview.chromium.org/6677069

git-svn-id: http://src.chromium.org/svn/trunk/src@78794 4ff67af0-8c30-449e-8e8b-ad334ec8d88c