Fix bug in shader and program managers.
The bug was that the code would search by id instead of by pointer.
This meant it was possible to free the wrong thing (ie, pass in
an info with id=4 and free a different info that also happened
to have id=4). This was most likely to happen when calling
MarkAsDeleted since MarkAsDelete sets the id to 0.

TEST=unit tests
BUG=75629

R=apatrick@chromium.org

Review URL: http://codereview.chromium.org/6721010

git-svn-id: http://src.chromium.org/svn/trunk/src@79130 4ff67af0-8c30-449e-8e8b-ad334ec8d88c