Die, DemuxerStream::QueryInterface, die!
This appears to have been an attempt at hand-rolling an RTTI system, all for the
use of a *single* class (!).  Removing it makes for cleaner, simpler, shorter
code.  It also enables delegating DemuxerStreams, which wasn't possible before
(b/c the overridable method was protected, so not visible to a HAS-A wrapper).

BUG=none
TEST=trybots

Review URL: http://codereview.chromium.org/6730006

git-svn-id: http://src.chromium.org/svn/trunk/src@79309 4ff67af0-8c30-449e-8e8b-ad334ec8d88c