Landing for Julien Tinnes, jln@google.com:
---
chroot to /proc instead of /tmp. This gets rid of a lot of unnecessary
complexity and fixes a race condition.

(Original idea from Markus)

The chroot helper will chroot to /proc/self/fdinfo (or /proc/self/fd). This is
pretty safe because access to this directory is protected by the ptrace() check
in the kernel and the helper is privileged.

Moreover, as soon as the helper _exit() and becomes a zombie, the directory
will be empty. Zygote should wait() for us to make everything deterministric.

We also export SBX_HELPER_PID so that Zygote can specifically wait for the
helper. 
---

BUG=76542
R=markus,agl
Review URL: http://codereview.chromium.org/6683056

git-svn-id: http://src.chromium.org/svn/trunk/src@79618 4ff67af0-8c30-449e-8e8b-ad334ec8d88c