ntp4: refactoring
pull out App and AppsPage into js classes. This cleans up the code and eventually new_tab.js should be a lot smaller and easier to read. This is especially motivated by wanting to do custom layout of the apps (so that we can do cool css animations between states, which browser native layout does not allow), which will be easier if we have an apps page object.

Note this comment from new_tab.css which mentions the problem I'm attempting to fix:
     TODO(rbyers): Should we take over responsibility for app layout ourself
     like the classic NTP's most-visited icons?  Or should we extend webkit
     somehow to support animation of the position of browser laid-out
     elements.

This
- ignores grabber code for now (totally breaks dragging)
- Doesn't use EventTracker (I think with a better JS object model it won't be necessary)
- gets rid of appContainer (I think it's superfluous if we are doing our own layout)
- removes some html templating (more of this to follow probably). I think it's redundant to have an html template, and then JS assertions to make sure the template looks the way we expect it to. Instead I made the dom hierarchy in question a sole responsibility of the javascript object.

I also removed the chrome web store "app" until I can ask Cole whether he wants to keep it (supporting it is annoying and weird since it's not re-order-able)

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6690063

git-svn-id: http://src.chromium.org/svn/trunk/src@80245 4ff67af0-8c30-449e-8e8b-ad334ec8d88c