ChromeFrame new window tests fail consistently on the IE8 builder because we expect the NewWindow3 event to be fired when a new window is opened. IE8 version 8.0.7601.17514 on the builder seems to be firing the NewWindow2 event consistently which breaks the expectations of these tests.
Added supports in the NewWindow2 handler in the event sink to call the OnNewBrowserWindow function
which most of the tests expect.

A limitation of this approach is we don't see the url in the NewWindow2 event. ChromeFrame tests
at this point don't seem to be using the url parameter.

BUG=none
TEST=ChromeFrame new window tests should pass on the IE8 builder.
TBR=amit
Review URL: http://codereview.chromium.org/6792004

git-svn-id: http://src.chromium.org/svn/trunk/src@80246 4ff67af0-8c30-449e-8e8b-ad334ec8d88c