Detach the PromoResourceService and WebResourceService from the Profile; have them store web resource data in local state instead.
We don't need multiple services for different Profiles; instead, there is a single service that captures data in local_state. The only preference stored in the Profile about the promo will be a bool indicating whether it's been closed or not, which will now be set outside the service.

Reviewers:
arv for general changes to PromoResourceService, & unittest.
erg for pulling the service out of the Profile and serving it up as a Singleton.
zmo for the changes I made to gpu_data_manager.
jstritar FYI, for changes to the PromoResourceService.


BUG=77155
TEST=promo resource service unittests.
Review URL: http://codereview.chromium.org/6736028

git-svn-id: http://src.chromium.org/svn/trunk/src@80318 4ff67af0-8c30-449e-8e8b-ad334ec8d88c