Fix crashing test
This is not a real crash, it is caused by wrapping TestSuite::Run in
an exception filter. DllRedirector test uses VerQueryValue to query
version from version resource. VerQueryValue, on windows XP, takes
a const pointer but tries to modify it causing access violation. 
So where do you fix this bug?  You guessed it right: in 
kernel32!UnhandledExceptionFilter :)
UnhandledExceptionfilter checks if the crash address belongs to a
resource section of a binary and then makes it writable. If we
install exception filter around VerQueryValue this does not happen
and we treat it as a crashed test.

solution is to remove the exception filter. This might result in
tests crashing levaing around zombie IE/Firefox. We will deal with
that when that happens.

TEST=DllRedirectorLoadingTest.TestDllRedirection
BUG=78209
Review URL: http://codereview.chromium.org/6792053

git-svn-id: http://src.chromium.org/svn/trunk/src@80623 4ff67af0-8c30-449e-8e8b-ad334ec8d88c