Drop RDH from ProfileImportProcessHost.
If a RDH is passed to the parent class, it will create a ResourceMessageFilter to handle resource loads.
ProfileImportProcessHost shouldn't need to load resources. It's just importing the profile.

This is part of the effort to pass a per-profile content::ResourceContext object to all message filters that need it. We need to avoid creating message filters unless we need them. We don't need a ResourceMessageFilter here.

BUG=78596
TEST=none


Review URL: http://codereview.chromium.org/6805006

git-svn-id: http://src.chromium.org/svn/trunk/src@80627 4ff67af0-8c30-449e-8e8b-ad334ec8d88c