WIP: Use inline "x" instead of "Remove" for languages in list.
This may be done, but I wonder if I need to remove the "Remove" button.

In order to make the language list items have the "X", I made LanguageList and LanguageListItem inherit from DeletableList and DeletableItem respectively.  When doing that, I found several css places which produced strangeness - the blue border when focusing the list, and the min-width:70px which made the "X" behave strangely on hover.  The blue border was identified in http://crbug.com/77730 and not removed in this patch.

While I was in there, I changed DeletableItem to produce an "li" tag instead of "div". This seemed to be more appropriate for what it was and seems to work, but could stand some review.

BUG=71495
TEST=browse to chrome://settings/languages, add & delete with "X"

Review URL: http://codereview.chromium.org/6759049

git-svn-id: http://src.chromium.org/svn/trunk/src@80683 4ff67af0-8c30-449e-8e8b-ad334ec8d88c