Reliability bot crash in PasswordStoreWin::GetLogins
Changes to keep all STL access on the same (DB) thread and add DCHECKS to assert this assumption.

Since the access to various STL containers is always done on the same thread,
this should be more reliable and not hit race conditions which corrupt the STL maps.

BUG=75296
TEST=chrome://settings/passwords works with no passwords.  unit_tests --gtest_filter=PasswordStore*.*

Review URL: http://codereview.chromium.org/6677125

git-svn-id: http://src.chromium.org/svn/trunk/src@80698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c