Define TransportTexture for 'sharing' texture between gpu and renderer
The purpose of TransportTexture is to allow creating textures in the gpu
process and be used in the renderer process. This is particularly useful
for hardware video decoder where the video frames are in the form of textures.
Creating these textures has to be originated from the GPU process, they are
then used in the renderer process for compositing.

This patch defines the class and has comments about its usage.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/6765020

git-svn-id: http://src.chromium.org/svn/trunk/src@80715 4ff67af0-8c30-449e-8e8b-ad334ec8d88c