Pepper3D API change - merging PPB_Surface3D and PPB_Context3D into PPB_Graphics3D.
The original API was designed to be similar to EGL a couple of reasons:
- EGL is a standard API that developers are familiar with and serves their purpose
- It would be relatively easy to write an EGL implementation on top of Pepper3D implementation to facilitate porting of existing EGL apps

However the restrictions in Pepper API do not allow certain use cases. The separation of context and surface allows using a single context to render to multiple instance or offscreen surfaces. But since each context is implicitly tied to a particular instance it is infeasible or confusing to draw to other instace surfaces. The offscreen use-case can be easily handled using Framebuffer objects (FBO). Writing an EGL wrapper for the new API should not be too difficult either.

Combining context and surface has advantages:
- Less complex API and implementation
- Convinient versioning

Please note that this is not ready to submit. I am only sending the proposed API. Once approved I will delete PPB_Surface3D and PPB_Context3D, and make other necessary changes to make sure everything compiles and works.
Review URL: http://codereview.chromium.org/6745010

git-svn-id: http://src.chromium.org/svn/trunk/src@80791 4ff67af0-8c30-449e-8e8b-ad334ec8d88c