New notification sent when extension uninstall is not allowed.
In the event that an extension uninstall is skipped (not allowed)
because that extension is not user-manageable, a new notification is
sent.  The automation hook UninstallExtensionById is revised to handle
this case, and a new PyAuto test is written to exercise this new
functionality (the test attempts to uninstall the WebStore and verifies that
this extension cannot be uninstalled).

BUG=76598
TEST=None.

Review URL: http://codereview.chromium.org/6794040

git-svn-id: http://src.chromium.org/svn/trunk/src@80814 4ff67af0-8c30-449e-8e8b-ad334ec8d88c