Remove unneeded v8 lockers from idb_bindings.cc.
Since this code is pulled into webkit_unit_tests, it is making us
need lockers there too, which stinks.

I don't see any reason for these to exist. My speculation is that
this code used to run in the browser process and was needed at
that time.

BUG=
TEST=

Review URL: http://codereview.chromium.org/6812038

git-svn-id: http://src.chromium.org/svn/trunk/src@80895 4ff67af0-8c30-449e-8e8b-ad334ec8d88c