Relanding http://src.chromium.org/viewvc/chrome?view=rev&revision=80938.
Add a new RenderViewHostObserver interface for filtering IPCs send to RenderViewHost. This is needed because with TabContentsObserver, an observer might not know which RenderViewHost a message came from duing a pending navigation. A side-benefit is that we only have to add these observers once, instead of remembering/knowing all the delegates that might want to filter these messages.

BUG=78629
Review URL: http://codereview.chromium.org/6813046

git-svn-id: http://src.chromium.org/svn/trunk/src@80984 4ff67af0-8c30-449e-8e8b-ad334ec8d88c