ChromeOS updater testing automation hooks that can be called from Python using PyAuto.
GetUpdateInfo() returns information about the updater, including what track (channel) the updater is on and what state the updater is in.
UpdateCheck() will initiate a check for updates and attempt to install one if there is one. It is a blocking call.
SetReleaseTrack() will set the release track to either "beta-channel" or "dev-channel".

BUG=chromium-os:12524
TEST=testSetReleaseTrack and testGetUpdateInfo included in chromeos_basic.py. UpdateCheck() tested manually, since an AU server or devserver is needed to test it.

Review URL: http://codereview.chromium.org/6813006

git-svn-id: http://src.chromium.org/svn/trunk/src@81129 4ff67af0-8c30-449e-8e8b-ad334ec8d88c