Wait for ProfileSyncService to be initialized on browser restart
The pyauto test sync.SyncTest.testRestartBrowser was flaky because, on
browser restart, the ProfileSyncService was sometimes in the middle of
being initialized when the test began its wait for sync completion.

This patch adds a check to make sure the sync service is initialized,
that will also wait for initialization. It also adds an extra check to
the sync.SyncTest.testRestartBrowser test case to make sure that
updates_received is 0 if a browser is restarted and nothing has changed
at the server end.

BUG=78677, 78548, 78863
TEST=pyauto test sync.SyncTest.testRestartBrowser

Review URL: http://codereview.chromium.org/6825023

git-svn-id: http://src.chromium.org/svn/trunk/src@81146 4ff67af0-8c30-449e-8e8b-ad334ec8d88c