GTK: Implement rough global menu support behind an about:flags flag.
This is a super early first cut of global menu support. It works by creating a GtkMenuBar inside chrome, attaching it to the widget hierarchy and then never showing it. This is sufficient for global menu support to detect the menu, at least on the Natty betas.

There's a lot to do before this is considered done. I haven't added Bookmark and History menus, which will need custom handling to keep state up to date with the main window. Some items may need to be moved from one menu to another to fit with Linux conventions.

Contains a bunch of new strings for global menu support. These need to go in before the M12 branch point. (I can iterate on the design afterwards; strings need translation.)

BUG=30213
TEST=Open a window on Unity. A minimal menu is in the menu bar.

Review URL: http://codereview.chromium.org/6812037

git-svn-id: http://src.chromium.org/svn/trunk/src@81170 4ff67af0-8c30-449e-8e8b-ad334ec8d88c