Add whitelist support to grit.
- Support -w path to add whitelists to grit (and grit_info).
- Use the whitelist to tag what items should be skipped when generating output.

A whitelist is used instead of a blacklist, since there has been some other
work into generating what resources are really needed by the build.

Another option instead of whitelists would have been to add yet more conditions
to the grd files, but they are already getting pretty complex with the
conditions already in the files.  And looking at trying to do conditions based
on features not only made the files more complex, but also doesn't handle the
problem when a single feature has some strings that are only valid for a subset
of the supported platforms.
Review URL: http://codereview.chromium.org/6821056

git-svn-id: http://src.chromium.org/svn/trunk/src@81393 4ff67af0-8c30-449e-8e8b-ad334ec8d88c