Add CORS options to URL requests. This information will be used by the URLLoader to create the AssociatedURLLoader that sends the request.
The CORS options are simplified to fit with what is already in PPAPI. For example, since we already have a trusted interface with universal access, setting the PP_URLREQUESTPROPERTY_ALLOWCROSSORIGINREQUESTS property TRUE corresponds to using access control to check the request. If we want to remove the trusted interface, then we should modify this CL so that we can specify Allow, Deny, or UseAccessControl. Also, I chose not to expose the SniffContent option from WebKit's loader. 
BUG=47354
TEST=test_shell_tests
Review URL: http://codereview.chromium.org/6755015

git-svn-id: http://src.chromium.org/svn/trunk/src@81604 4ff67af0-8c30-449e-8e8b-ad334ec8d88c