History: get thumbnails from temporary in-memory storage
This is a workaround for history stack and tabsidebar.
Also query thumbnails from a temporary in-memory storage. This
can make part of work for history stack and can make tabsidebar
work well. For history stack, since the size of in-memory storage
is 20, there is possible blank thumbnail when temporary thumbnails
are cleaned up.

Also this can't work for bookmark manager if you restart browser.