Split OSMesa implementations of *GLContext into GLContextOSMesa and *GLSurfaceOSMesa.
Surfaces are independent of contexts in GL. To facilitate sharing of surfaces between processes, I have separated them from the notion of contexts because contexts cannot be shared between processes.

I started with EGL in r81512 and WGL in r81807. This is the same thing for OSMesa.

GLContextOSMesa still has a pointer to a surface and still has some surface specific operations that just forward through to it. Once I have refactored all the GLContext implementations in this way, I will remove these pointers and the surface specific opertations.

There will not be "view" and "offscreen" GL contexts. Rather there will be a single context type for each backend which can be made current with a surface that directs output either to a view or offscreen surface.

TEST=WebGL locally, try
BUG=none
Review URL: http://codereview.chromium.org/6864031

git-svn-id: http://src.chromium.org/svn/trunk/src@81998 4ff67af0-8c30-449e-8e8b-ad334ec8d88c