RWHV: Enable popup suggestion item's selection (BMC# 15837)
We don't need to support double click or pinch gesture for popup window,
So we need to by pass the mouse release event delay routine. Otherwise
the item in popup window will not be selectable, cause the launcher
will focus out the on foucs widget before the item is selected.