Text selection: make it work in input text field
We can't select text in input text fields if the page can be panned
in horizontal. The solution is to add one flag to track the status
and forbids the pan in horizontal in this kind of selection.

Another fix is to make flick/pan work in text selection mode. This
is fixed by setting correct flag in the mouse press/release event
handler. When doing modifing selection, disable interactive for
viewport when mouse press and enable it again when mouse release.