Touch Events: pass through touch events to JS when needed
W3C draft proposed touch events support in JS but currently there
is no 'touchmove' event passed through to JS.

Now 'touchstart' and 'touchend' is directly forwarded but for
'touchmove', the solution is to pass through it if and only if the
target node has touch event listener.

There are many problems here when co-working with other touch and
gestures. But the basic function is workable. You can test it with
below url:
http://o.sitepen.com/labs/code/iphone/events.html