Merged and tested r84353 from trunk:
Thanks to piman, added necessary glFlush to the IOSurface code path on
the Mac. This was the root cause of the delay in transmission of
rendering results from the GPU process to the browser process.

Undid the changes from the fix for Issue 80602; they are no longer
needed.

Reran various test cases on Mac; no regressions.

BUG=81740
TEST=none
TBR=apatrick,piman
Review URL: http://codereview.chromium.org/6932062

git-svn-id: http://src.chromium.org/svn/branches/742/src@84379 4ff67af0-8c30-449e-8e8b-ad334ec8d88c