Bug fix: render error of news.baidu.com and www.sohu.com.
The root cause might be that browser service will take thumbnail
of tab when tab status changes. But the PaintAtSize of render widget
which is backend of thumbnail service will resize web widget.
When page is loading, while flick to scroll, this issue is easy to happen.

Fix it by change PaintAtSize to avoid resizing web widget.

Test:
1. navigate news.baidu.com and flick to scroll.
2. same as www.sohu.com
3. make sure thumbnails take are right (~/.config/internet-panel/thumbnails)
4. pinch to zoom works