UX: Refine bookmark bar
There are some gaps between current implemention and  visual design,
inculding title font size, title font family, background image and
some left and right margin. And this commit has solved these problems.

Maybe there are some improvements in the future, such as the bookmark item can't be
highlight when it is pressed, and it is more charming that the width of
item is according with the length of title.