Merge 86141 - This adds a Flush to the session backend to help with data corruption at ChromeOS shutdown.
We're seeing crashes due to corrupted history items in the session data on ChromeOS, and this is a bandaid to help reduce them.  It's not the real solution, which is to have more corruption-robust history item storage and recovery, but it shouldn't cause any harm.

Our supposition here is that corruption is occurring because we shutdown while session data hasn't been flushed to the disk, thus corrupting the history data, which is fragile, and we crash the next time we read in the session.

This CL will help because we flush after appending system data.  Other CLs will help to clean up the bad history data, and be more robust in the case of corruption.

BUG=chromium-os:15478, chromium-os:15350
TEST=Loaded existing current state, exited, reloaded, ran unit test.

Review URL: http://codereview.chromium.org/7044018

TBR=zelidrag@google.com
Review URL: http://codereview.chromium.org/7063027

git-svn-id: http://src.chromium.org/svn/branches/742/src@86506 4ff67af0-8c30-449e-8e8b-ad334ec8d88c