Merge 87082 - Several fixes to the UI for enabling SIM-locked modems.
- Clicking "Enable Mobile Data" from either the network dropdown
menu or the internet settings page results in the SIM unlock dialog
being posted if the SIM is locked.
- If the Enable button on the settings page was pressed, and the SIM
unlock dialog is cancel, the Enable button, which had been disabled
(greyed out) when clicked, is re-enabled.
- If the correct PIN (or PUK) is entered, and the SIM becomes unlocked,
we now perform the actual device enable operation.

Also changed the definitions of the PIN error messages, to match an
about-to-be-committed change made on the flimflam side.

BUG=chromium-os-partner:3538
TEST=Press enable on the settings page, bringing up unlock dialog.
- Cancel the dialog. See that the Enable button is active.
- Enter an incorrect PIN and see that it behaves as before, i.e.,
the dialog stays up and reports the new retries-allowed count.
- Enter the correct PIN and see that the device becomes enabled, and
the button changes to "Disable Mobile Data".

Try all three of the above scenarios using the Enable choice in
the network dropdown menu, with and without any Chrome tab displaying
the settings page. Make sure the results are the same as they were when
the Enable button on the settings page was used (except of course,
the Enable button won't be displayed, so the part where you check
that it's active after cancelling doesn't apply.

R=stevenjb@chromium.org,nkostylev@chromium.org,zelidrag@chromium.org

Review URL: http://codereview.chromium.org/7075015

TBR=ers@chromium.org
Review URL: http://codereview.chromium.org/7078011

git-svn-id: http://src.chromium.org/svn/branches/742/src@87087 4ff67af0-8c30-449e-8e8b-ad334ec8d88c