Bug#15923 Browser app asks user to make it the default browser
If the meego-app-browser is not the default browser, the related info
bar will be poped up, displaying that set meego-app-browser to be
default browser or don't ask me again information. In the
/usr/bin/meefo-app-browser shell script, the meego-app-browser.desktop
has already been setten to be default browser, but the function
ShellIntegration::IsDefaultBrowser() use the command "xdg-setting check
default-web-browser chromium-devel.desktop", which is the root cause.