UX: bring selection handler to front end -- browser part
Previously selection handler is drawn in Webkit with
a circle. It is not consistent with other selection handlers which
use a themed image.

This commit is to replace Webkit circle with the same themed image with
other components, which needs to expose information such as selection
position, height etc to browser and then to QML part.

reviewed by Junmin