Merge 87732 - file manager: Fix okButton updating.
update the okButton whenever the directory changes.
This fixes bug 15845, which was caused by trying to update the button before we had set the current directory, and bug 15464, which was due to the fact that we relied on "Selection changed" events to trigger button updates.

BUG=chromium-os:15845,chromium-os:15464
TEST=manual testing on chromeos-chrome on linux

Review URL: http://codereview.chromium.org/7109033

TBR=rginda@chromium.org
Review URL: http://codereview.chromium.org/7054047

git-svn-id: http://src.chromium.org/svn/branches/742/src@87798 4ff67af0-8c30-449e-8e8b-ad334ec8d88c