Render: always send preferred size no matter whether it's changed
SetPreferredSize will check whether the new SetPrerredSize is
equal to the last setting. If yes, don't send IPC to render process.
But it's not suitable since when the render process crashes and a new
render process is created, preferred size won't be sent to the
new render process.

Fix this by always sending preferred size.
Reviewed by Ningxin.