Omnibox: refactor select feature in Omnibox [BMC#18808]
[issues]
1. Previously, when handling selectall signal, use shouldSelectAll to
indicate in selectAll mode, if url text is not selected all
re-selectall. That's because after focus in event immediately there is a
mouse event which will clear the selection. So the selectAll logic is
complicated at present when adding more feature in Omnibox.

2. BMC#18808 hard to move cursor in Omnibox. This because height of
TextInput is determined by the font size in Omnibox. TextInput doesn't
fill fully in Omnibox. There are gaps above and under TextInput. If
touch these gaps, no action would be done.

[solutions]
1. For issue 1: Use a timer to delay selectAll which is more natural than re-select
text many times for one selectAll signal.
2. For issue 2: pass mouse event outside urlTextInput in Omnibox to urlTextInput

Reviewed by Yin Li