UX Fix the issue of popping up two crash dialogs
When the render process is killed or crashed, sometimes there are two
dialogs popped up. The root cause is that some special web pages will
popup some windows such as www.sina.com.cn, and there are two
tab_contents objects will be created in the browser, whether the popup
window option is set to be allowed or blocked. So, if we set the popup
window option to be blocked, the popup window can't be diaplayed in the
screen, but it still can receive the crashed message, when the render
process is killed or crashed.

The solution is that we need check the window is blocked or not, before
add the dialog into queue.

Test case:
1.www.sina.com.cn
2.www.amazon.com
3.www.meego.com
4.and so on

Reviewed by James[james.wei@intel.com]