Revert "BMC #15524 : Enable right click for windowless flash plugin"
The flash plugin still get hang when receiving right mouse event when
running in windowless mode.

We don't encounter this before because we delay the mouse press event
on normal routing, while we will encounter this issue when we fake a
right mouse event from tapandhold gesture, in which case it don't get
delayed. need to wait for flash plugin to fix this issue before we
re-enable it.

This reverts commit dd20ca2d41d4197d5a68a4e29649418941c8f044.

Reviewed by Ningxin