Tiling: Adjust minimal scale factor in portrait mode
In portrait mode, the content width of web page is possible to
be larger than the width of screen. This fix is to adjust the
minimal scale factor dynamically according to the content width
and screen width.

Once orientation is changed from landscape to portrait mode, the
content scale is designed to not be adjusted, but the minimal scale
factor for scaling to screen is calculated. User is allowed to
zoom out web page to view its full picture by pinch gesture.

Test case:
1. open www.cnn.com in landscape mode
2. switch to portrait
3. Pinch to zoom out web page
4. Full picture of cnn.com is showed

Reviewed by Yongsheng.