Fix BMC 21271 tabListUpdated signal not emited when user create a new tab.
Refined the browser APIs. Instead of taking thumbnails and emiting tabInfoUpdated
signals in TabChangedAt() method, these actions will happen when a page is seleted,
the title of a page updated or the browser became minimized, so as to reduce the
fequency of taking thumbnails and firing signals. Details:

1. When current page is going to be switched to another one, the thumbnail and other
information(title, url etc.) of current page will be taken.
2. When open a new tab or close a opened tab, the tabListUpdated signal will be emitted.
3. Add a "refreshTabList" DBus method to update the information of all tabs, a
tabListUpdated signal will be emitted after call this API.
4. Once a thumbnail and other information have been taken after the page is completely
loaded, these infomation would not be refreshed any more.

Reviewed by Hongbo.