Fix BMC 21404 there are many highlight rectangles in the process of page scroll
The link would be highlighted when the link is pressed. But if you pan a
page from the position of the link, the highlight rectangle cannot be
disappeared after releasing your finger. Repeat this step, there are
many highlight rectangles in the page finally.

The root cause is that flickable steals mouse event during pan a page.
To workaround this, listen the movementStarted event in flickable, if
received this event, generate a mouse release event and forward it.

Reviewed by Hongbo.